NXT-992 Move acceptCookieBanner to support/action and stop it failing tests when the civic banner never loads
