+++ functional-tests/steps/given.ts
@@ -4,31 +4,12 @@ import { expect } from "expect-webdriverio";
 import { clickElement } from "@nice-digital/wdio-cucumber-steps/lib/support/action/clickElement.js";
 import { waitForDisplayed } from "@nice-digital/wdio-cucumber-steps/lib/support/action/waitForDisplayed.js";
 
+import { acceptCookieBanner } from "../support/action/acceptCookieBanner.js";
 import {
 	indicatorFilter,
 	indicatorFilterDate,
 } from "../support/action/indicatorFilter.js";
 import { getPath, PageName } from "../support/pagePaths.js";
-
-/**
- * Accept all cookies using the NICE cookie banner.
- * We use this as the cookie banner blocks clicks.
- */
-export async function acceptCookieBanner(): Promise<void> {
-	// The cookie banner (and license key) load async from the CDN
-	// so we need to wait for it to load
-	const cookieBannerElement = await $("body #ccc");
-	await cookieBannerElement.waitForExist({ timeout: 2000 });
-
-	const acceptCookiesButtonElement = await cookieBannerElement.$(
-		"button.ccc-accept-button"
-	);
-
-	// If cookies have already been chosen then the accept button doesn't show
-	if (await acceptCookiesButtonElement.isDisplayed()) {
-		await acceptCookiesButtonElement.click();
-	}
-}
 
 /*! https://github.com/webdriverio/cucumber-boilerplate/blob/master/src/support/action/openWebsite.js */
 /**
